requirements.txt
altair==5.2.0
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
crispy-bootstrap5==2023.10
cycler==0.12.1
dj-database-url==2.1.0
Django==5.0.2
django-crispy-forms==2.1
django-heroku==0.3.1
django-leaflet==0.29.0
djangorestframework==3.14.0
docopt==0.6.2
fonttools==4.48.1
GDAL @ file:///C:/Users/opote/pipwin/GDAL-3.4.3-cp311-cp311-win_amd64.whl#sha256=f78861fb5115d5c2f8cf3c52a492ff548da9e1256dc84088947379f90e77e5b6
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.41
gunicorn==21.2.0
idna==3.6
import-export==0.3.1
importlib-metadata==7.0.1
Jinja2==3.1.3
Js2Py==0.74
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mdurl==0.1.2
numerize==0.12
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pipwin==0.5.2
plotly==5.18.0
protobuf==4.25.2
psycopg2==2.9.9
pyarrow==15.0.0
pydeck==0.8.1b0
Pygments==2.17.2
pyjsparser==2.7.1
pyparsing==3.1.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2024.1
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sqlparse==0.4.4
streamlit==1.31.0
streamlit-option-menu==0.3.12
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
validators==0.22.0
watchdog==4.0.0
whitenoise==6.6.0
zipp==3.17.0
